move api.v7 test
