+++ api.v7/kodo/operator_test.go
@@ -1,4 +1,4 @@
-package test
+package kodo
 
 import (
 	"bytes"
